Move router to an individual file
